==> doc/01-processo.Rmd <==
---
title: "Processo"
subtitle: "UPA415"
author: "christian.scarabelli@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ../misc/custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
```



# Processo

Nel corso degli anni 2021 e 2022, il collaudo automatico UPA su barre pelate della marca interna 415M3 ha messo in evidenza colate con deviati alti per difetto centrale. Questo risulta essere problematico per l’aggravio costi causati dagli scarti extra-standard e per i possibili ritardi nelle consegne ai clienti.
A seguito di tali eventi si vogliono analizzare i dati di processo. In questo specifico caso, si lavora su una popolazione di dati che fa riferimento alle **BARRE PELATE**.
I dati di processo raccolti sono relativi alle analisi chimiche finali (estratti da APC *RCCHIM01*) ed ai dati di colaggio in colata continua (estratti da Qlik *Danieli Colata Continua*). Verranno pertanto **trascurati sia i parametri di processo di fusione-fabbricazione sia quelli a valle del colaggio in CCO, perché ritenuti ininfluenti sulle variabili di risposta**.
I dati di deviato al UPA (in questo caso UPA6) sono stati estratti da APC (*RPROC03*) e fanno riferimento (come sommatoria [#]) alle colate testate.  *Consideriamo, per semplificare, il dato omogeneo sulle varie schede della colata*.
Per riassumere, i dati di processo (103 colate totali) sono:

 + Analisi chimica finale per ogni colata;
 + Dati di processo di colata continua per ogni colata;
 + Dati di deviato (SC+DEV) al UPA per difetto centrale per ogni colata.
 
I dati fanno riferimento agli anni 2021 e 2022.
 
[#] per sommatoria si indente che i dati sono stati pivotati per odp, tenendo conto del fatto che impiegato e scarto/deviato sono su due righe differenti.
