==> doc/05-data-discovery.Rmd <==
---
title: "Data Discovery "
subtitle: "CRO304"
author: "pbrunier@cogne.com"
version: 1.0 
date: "`r Sys.Date()`"
output:
  rmdformats::readthedown:
    css: ../misc/custom.css
    df_print: paged
    gallery: no
    highlight: default
    html_document: null
    lightbox: yes
    number_sections: yes
    self_contained: yes
    thumbnails: no
editor_options:
  chunk_output_type: console
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = FALSE,
                      warning = FALSE,
                      message = FALSE,
                      error = FALSE,
                      fig.height = 7,
                      fig.width = 10,
                      collapse = TRUE,
                      cols.print=20)
rm(list = ls(all.names = TRUE))
require(dplyr)
require(rlang)
require(kableExtra)
require(readr)
require(ggplot2)
require(reticulate)
require(rpart)
require(rpart.plot)
require(tidyr)
require(randomForest)
source('~/dev/UPA415/R/function.R')
```


leggiamo i dati 

```{r loading_data_R}
upa415 <- read_rds('/data/UPA415/upa415.rds')
                       
```

```{python loading_data_python}

import pandas as pd

df = pd.read_pickle('/data/UPA415/upa415.pkl')
                       
```

# Analisi temporale dei deviati


```{python analisi deviati}
import matplotlib.pyplot as plt
from scipy.interpolate import interp1d
import statsmodels.api as sm

colori = ['blue','limegreen','darkorange','magenta','dodgerblue','dimgray','darkviolet']

# genero gli scarti percentuali


df['y_perc'] = df.y * 100
df['sc_fbh_07_perc'] = (df.fbh_07_scarto + df.fbh_07_dev) / df.fbh_07_imp * 100
df['sc_fbh_10_perc'] = (df.fbh_10_scarto + df.fbh_10_dev) / df.fbh_10_imp * 100
df['sc_fbh_15_perc'] = (df.fbh_15_scarto + df.fbh_15_dev) / df.fbh_15_imp * 100



lowess = sm.nonparametric.lowess(df.y_perc,df.index, frac=.1)

# unpack the lowess smoothed points to their values
lowess_x = list(zip(*lowess))[0]
lowess_y = list(zip(*lowess))[1]

plt.style.use('seaborn')

plt.figure()
plt.plot(df.index,df.y_perc,'--o',linewidth=0.5,markersize=3,label='data')
plt.plot(lowess_x,lowess_y,'crimson',linewidth=1,label='lowess')
plt.xlabel('data')
plt.ylabel('scarto percentuale')
plt.title('andamento nel tempo dello scarto totale')
plt.legend(loc='best')
plt.show()

list_deviati = ['sc_fbh_07_perc','sc_fbh_10_perc','sc_fbh_15_perc']

i = 0
for ll in list_deviati:
  plt.figure()

  plt.plot(df.index,df[ll],'--v',color=colori[i],
  linewidth=0.5,label=ll,markersize=2)
  
  lowess = sm.nonparametric.lowess(df[ll],df.index, frac=.6)

  # unpack the lowess smoothed points to their values
  lowess_x = list(zip(*lowess))[0]
  lowess_y = list(zip(*lowess))[1]

  plt.plot(lowess_x,lowess_y,color=colori[i],linewidth=1,label='lowess')
  i += 1

  plt.legend(loc='best')
  plt.xlabel('data')
  plt.ylabel('%s' %ll)
  plt.title('andamento nel tempo per %s' %ll)
  plt.show()
plt.style.use('default')

```

# analisi delle variabili di input nel tempo

```{r analisi_ivso, result = 'asis'}
dati <- py$df
# aggiungo colonna index (tidyverse)
dati <- tibble::rowid_to_column(dati, "index")

vars <- dati %>%
  select(C:bassa_press_media_reale) %>%
  names()
titles <- paste('## Andamento nel tempo della variabile ', vars)

n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  title_i <- titles[i]
  pl <- ggplot(dati)+
    geom_line(aes(index,!!sym(var_i)))+
                geom_smooth(aes(index,!!sym(var_i)),method = 'loess',
                            se=FALSE, col='red')
              
              cat(title_i, '\n')
              print(pl)
              cat('<p>')
}                     
```

# Analisi delle correlazioni variabili input VS variabile risposta (scarto percentuale)

analisi di correlazioni

```{r correlazioni,results='asis'}
vars <- dati %>%
  select(C:bassa_press_media_reale) %>%
  names()
titles <- paste('## Andamento nel tempo della variabile ', vars)

n <- length(vars)

for (i in seq_len(n)) {
  var_i <- vars[i]
  title_i <- titles[i]
  pl <- ggplot(dati)+
    geom_line(aes(!!sym(var_i),y_perc))+
                geom_smooth(aes(!!sym(var_i),y_perc),
                            method ='loess',
                            se=FALSE, col='red')
              
              cat(title_i, '\n')
              print(pl)
              cat('<p>')
}    
```

# decision tree

```{r decision_tree}
dati_tree <- dati %>% 
  select(y_perc,C:bassa_press_media_reale)

fm <- rpart(formula= y_perc ~ .,
            data = dati_tree,
            method = 'anova',
            weights = dati$imp, 
            cp = 0.02)

rpart.plot(fm,
           extra=101)


```

# random forest

```{r randomforest}

dati_tree <- na.omit(dati_tree)
rf <- randomForest(formula = y_perc ~ .,
                 data = dati_tree,
                 weights = dati$imp,
                 ntree = 1000)

# Variable importance plot
varImpPlot(rf, main='Variable importance Plot')
# Plotting model
plot(rf,main='convergence plot')
# Importance plot
importance(rf)
```
